Signature matrix creation with DWLS using genes identified by a differential analysis

Signature matrix creation with DWLS using genes identified by a differential analysis

build_model_dwls(
  single_cell_object,
  cell_type_annotations,
  dwls_method = c("mast", "seurat", "mast_optimized"),
  path = NULL,
  verbose = FALSE,
  diff_cutoff = 0.5,
  pval_cutoff = 0.01,
  ncores = 1
)

Arguments

single_cell_object

A matrix with the single-cell data. Rows are genes, columns are samples. Row and column names need to be set.

cell_type_annotations

A vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object.

dwls_method

The method used to create the signature matrix. Options are "mast", "mast_optimized" and "seurat"

path

The path where the generated files will be saved. If path=NULL, the generated files will be discarded.

verbose

Whether to produce an output on the console.

diff_cutoff

Cutoff to determine the FC-limit. How low can the lowest fold change be to still be considered differentially expressed?

pval_cutoff

Cutoff to determine the pVal-limit. How high can the highest p-Value be to still be considered statistically significant?

ncores

Choose how many core to use for signature calculation. Is only applied when using 'mast' as method.

Value

The signature matrix. Rows are genes, columns are cell types.