Builds Scaden Model from scRNA data

The model is saved in a defined directory.

build_model_scaden(
  single_cell_object,
  cell_type_annotations,
  bulk_gene_expression,
  model_path = NULL,
  temp_dir = NULL,
  batch_size = 128,
  learning_rate = 1e-04,
  steps = 5000,
  var_cutoff = NULL,
  cells = 100,
  samples = 1000,
  dataset_name = "scaden",
  verbose = FALSE
)

Arguments

single_cell_object

A matrix with the single-cell data. Rows are genes, columns are samples. Row and column names need to be set.

cell_type_annotations

A vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object.

bulk_gene_expression

A matrix of bulk data. Rows are genes, columns are samples. Row and column names need to be set.

model_path

Path where model directory should be created (optional).

temp_dir

The temporary directory to use for the computations (optional)

batch_size

Training of model: Batch size to use for training (default: 128).

learning_rate

Training of model: Learning rate used for training (default: 1E-4).

steps

Training of model: Number of training steps (default: 5000).

var_cutoff

Training data processing: Filter out genes with a variance less than the specified cutoff. A low cutoff is recommended,this should only remove genes that are obviously uninformative. (default NULL).

cells

Bulk simulation: Number of cells per sample (default: 100).

samples

Bulk simulation: Number of samples to simulate (default: 1000).

dataset_name

Bulk simulation: Name of simulated dataset (default scaden).

verbose

Whether to produce an output on the console (default: false).

Value

The path to the scaden model.