Signature matrix creation with CIBERSORTx

Signature matrix creation with CIBERSORTx

build_model_cibersortx(
  single_cell_object,
  cell_type_annotations,
  verbose = FALSE,
  input_dir = NULL,
  output_dir = NULL,
  display_heatmap = FALSE,
  k_max = 999,
  ...
)

Arguments

single_cell_object
A matrix with the single-cell data. Rows are genes, columns are samples. Row and column names need to be set.

cell_type_annotations
A vector of the cell type annotations. Has to be in the same order as the samples in single_cell_object.

verbose
Whether to produce an output on the console.

input_dir
The directory in which the input files can be found (or are created in). Default is a temporary directory.

output_dir
The directory in which the output files are saved. Default is a temporary directory.

display_heatmap
Whether to display the generated heatmap.

k_max
Maximum condition number (default: 999). Will be added to the file name.

...
Additional parameters supplied to the algorithm. Options are:

g_min <int> Minimum number of genes per cell type in sig. matrix (default: 300).

g_max <int> Maximum number of genes per cell type in sig. matrix (default: 500).

q_value <int> Q-value threshold for differential expression (default: 0.01).

filter <bool> Remove non-hematopoietic genes (default: FALSE).

remake <bool> Remake signature gene matrix (default: False).

replicates <int> Number of replicates to use for building scRNAseq reference file (default: 5).

sampling <float> Fraction of available single cell GEPs selected using random sampling (default: 0.5).

fraction <float> Fraction of cells of same identity showing evidence of expression (default: 0.75).

Value

The signature matrix. Rows are genes, columns are cell types.